+++ lib/matt/support/puts.rb
@@ -21,6 +21,20 @@ module Matt
         configuration.stderr.puts(*args, &bl)
       end
 
+      def info(message)
+        message = "#{self.to_s.ljust(30)} -- #{message}"
+        return $stderr.puts(message) unless configuration
+        return unless configuration.debug_level >= Configuration::DEBUG_INFO
+        puts_err(message)
+      end
+
+      def debug(message)
+        message = "#{self.to_s.ljust(30)} -- #{message}"
+        return $stderr.puts(message) unless configuration
+        return unless configuration.debug_level >= Configuration::DEBUG_VERBOSE
+        puts_err(message)
+      end
+
     end # module Puts
   end # module Support
 end # module Matt