+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Added `Matt.env!` to use an env variable that must exist.
 
+* Added --silent and --verbose options for various debug levels.
+
 ## 1.0.0 - 2021-06-09
 
 * First official version used in production.